ai_engine/requirements.txt

# Core
python-dotenv==1.0.0
pydantic==2.5.0

# LLM
openai==1.6.0
anthropic==0.8.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langchain-upstage==0.7.4
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
llama-index==0.9.30

# Vector DB
qdrant-client==1.7.0

# Knowledge Graph
neo4j==5.14.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# ML & NLP
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0

# Utils
tenacity==8.2.3

